+++ lib/rubocop/cop/commissioner.rb
@@ -34,7 +34,7 @@ module RuboCop
 
       callback_methods.each do |callback|
         next unless method_defined?(callback)
-        class_eval <<-RUBY, __FILE__, __LINE__ + 1
+        class_eval(<<-RUBY, __FILE__, __LINE__ + 1)
           def #{callback}(node)
             @callbacks[:"#{callback}"] ||= @cops.select do |cop|
               cop.respond_to?(:"#{callback}")